Refactor: binding.utils renamed to route.bindings and moved to /router
+ Updates import declaration in Resource.vue
